完善驗證登入，增加學生資料刪除提示

+++ routes/courses.js
@@ -8,7 +8,7 @@ const router = express.Router();
 // 首頁
 
 router.get('/', (req, res, next) => {
-    // if(req.session.userName){
+    if(req.session.userName){
         Course.find((err, docs) => {
             //排序
             let sortcourse = docs.sort((a,b) =>{
@@ -20,49 +20,61 @@ router.get('/', (req, res, next) => {
             });
             
         });
-    // }else{
-    //     req.flash('error','無法查看課程，請先登入')
-    //     res.redirect('login');
-    // }
+    }else{
+        req.flash('error','無法查看課程，請先登入')
+        res.redirect('login');
+    }
 });
 
 // 跳轉新增課程頁面
 router.get('/add', (req, res, next) =>  {
+    if(req.session.userName){
         Course.find((err, docs) => {
             res.render('add', {
                 title: '新增課程',
                 courses: docs
             });
         });
-    
+    }else{
+        res.redirect('/courses');
+    }
 });
 
 // 新增課程
 router.post('/add', (req, res, next) =>  {
-    let courses = new Course({
-        course: req.body.course,
-        date: req.body.date,
-        lecturer: req.body.lecturer
-    });
-    courses.save((err, doc) => {
-        res.redirect('/courses');
-    });
+    if(req.session.userName){
+        let courses = new Course({
+            course: req.body.course,
+            date: req.body.date,
+            lecturer: req.body.lecturer
+        });
+        courses.save((err, doc) => {
+            res.redirect('/courses');
+        });
+    }else{
+        res.redirect('/');
+    }
     
 });
 
 // 根據id刪除對應課程資料
 router.get('/del', (req, res, next) => {
-    let id = req.query.id;
-    if (id && id != '') {
-        Course.findByIdAndRemove(id, (err, docs) => {
-            res.redirect('/courses');
-        });
+    if(req.session.userName){
+        let id = req.query.id;
+        if (id && id != '') {
+            Course.findByIdAndRemove(id, (err, docs) => {
+                res.redirect('/courses');
+            });
+        }
+    }else{
+        res.redirect('/');
     }
     
 });
 
 // 查詢對應修改記錄，並跳轉至修改頁面
 router.get('/update', (req, res, next) => {
+    if(req.session.userName){
         let id = req.query.id;
         if (id && id != '') {
             Course.findById(id, (err, docs) => {
@@ -72,25 +84,31 @@ router.get('/update', (req, res, next) => {
                 });
             });
         }
+    }else{
+        res.redirect('/courses');
+    }
     
     
 });
 
 // 修改課程資料
 router.post('/update', (req, res, next) =>  {
-    
-    let courses = {
-        course: req.body.course,
-        date: req.body.date,
-        lecturer: req.body.lecturer
-    };
+    if(req.session.userName){
+        let courses = {
+            course: req.body.course,
+            date: req.body.date,
+            lecturer: req.body.lecturer
+        };
 
-    let id = req.body.id;
+        let id = req.body.id;
 
-    if (id && id != '') {
-        Course.findByIdAndUpdate(id, courses, (err, docs) => {
-            res.redirect('/courses');
-        });
+        if (id && id != '') {
+            Course.findByIdAndUpdate(id, courses, (err, docs) => {
+                res.redirect('/courses');
+            });
+        }
+    }else{
+        res.redirect('/');
     }
     
 });
